refactor: 수익률 계산하는 기능 수정
- 수익률 소수점 둘째 자리에서 반올림
- 수익률 Double 로 수정
- LottoResultDetail에 소수점 둘째자리에서 반올림한 수익률 프로퍼티 추가
(Application 에서 임시 출력 코드 추가)

+++ src/main/kotlin/lotto/LottoResult.kt
@@ -41,11 +41,11 @@ class LottoResult(
         return false
     }
 
-    private fun getRateOfReturn(lottoRankList: List<LottoRank>): Float {
-        // (최종 수익 / 구입 금액) * 100
+    // 최종 수익률: (최종 수익 / 구입 금액) * 100
+    private fun getRateOfReturn(lottoRankList: List<LottoRank>): Double {
         val purchasePrice = lottoList.size * LottoConstants.PRICE
         val finalProfit = lottoRankList.sumOf { it.price }
-        val rateOfReturn = (finalProfit / purchasePrice).toFloat() * PERCENTAGE_MULTIPLIER
+        val rateOfReturn = (finalProfit / purchasePrice).toDouble() * PERCENTAGE_MULTIPLIER
         return rateOfReturn
     }
